move some components around
