+++ redux-form-tutorial/src/main/typescript/client.tsx
@@ -3,7 +3,8 @@ import * as ReactDOM from "react-dom"
 import {Provider} from "react-redux"
 import store from "./store"
 import Header from "./components/Header"
-import {SimpleForm} from "./components/SimpleForm/SimpleForm"
+import UsersList from "./components/UsersList"
+import DemoForm from "./components/DemoForm"
 import "../resources/css/styling"
 import 'react-datepicker/dist/react-datepicker-cssmodules.css'
 
@@ -11,7 +12,8 @@ ReactDOM.render(
     <Provider store={store}>
         <div>
             <Header/>
-            <SimpleForm/>
+            <UsersList/>
+            <DemoForm/>
         </div>
     </Provider>,
     document.getElementById('app'),
