--- redux-form-tutorial/src/main/typescript/components/SimpleForm/SimpleForm.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react'
-import DemoForm from "./DemoForm"
-import UsersList from "./UsersList"
-
-export const SimpleForm = () => <div><UsersList/><DemoForm/></div>
